pyproject.toml
[tool.poetry]
name = "sherlock-wind"
version = "0.1.0"
description = ""
authors = ["jthetzel <jthetzel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.63.0"}
cdsapi = "^0.5.1"
python-dotenv = "^0.15.0"
smart-open = {extras = ["all"], version = "^4.2.0"}
jupyterlab = "^3.0.9"
httpx = "^0.16.1"
xarray = "^0.17.0"
Pydap = "^3.2.2"
scipy = "^1.6.1"
zarr = "^2.6.1"
fsspec = "^0.8.7"
gcsfs = "^0.7.2"
matplotlib = "^3.3.4"
Cartopy = "^0.18.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
